Add files via upload
To create a simple chatting application with one server and two clients connected in Java, you can follow the steps below: Create a Server class that listens for incoming client connections and handles communication between the clients. The server should be able to accept multiple client connections and broadcast messages from one client to all other connected clients. Create a ClientHandler class that handles communication between the server and a single client. The ClientHandler class should be able to receive messages from the client and send them to the server, as well as receive messages from the server and send them to the client. Create a Client class that connects to the server and sends/receives messages. The Client class should be able to display incoming messages and allow the user to enter outgoing messages. Implement the Server, ClientHandler, and Client classes using Java sockets and input/output streams.